refactor :: move auth resolver package
