+++ spaced/src/main/java/com/dnd/spaced/domain/auth/presentation/AuthController.java
@@ -8,8 +8,8 @@ import com.dnd.spaced.domain.auth.presentation.dto.response.TokenResponse;
 import com.dnd.spaced.domain.auth.presentation.exception.RefreshTokenNotFoundException;
 import com.dnd.spaced.global.config.properties.TokenProperties;
 import com.dnd.spaced.global.controller.ResponseEntityConst;
-import com.dnd.spaced.global.resolver.AuthAccount;
-import com.dnd.spaced.global.resolver.AuthAccountInfo;
+import com.dnd.spaced.global.resolver.auth.AuthAccount;
+import com.dnd.spaced.global.resolver.auth.AuthAccountInfo;
 import com.dnd.spaced.global.security.handler.OAuth2SuccessHandler;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
