+++ spaced/src/main/java/com/dnd/spaced/global/interceptor/AuthInterceptor.java
@@ -3,8 +3,8 @@ package com.dnd.spaced.global.interceptor;
 import com.dnd.spaced.domain.account.application.ValidateAccountService;
 import com.dnd.spaced.domain.auth.application.exception.ForbiddenAccountException;
 import com.dnd.spaced.domain.auth.domain.TokenDecoder;
-import com.dnd.spaced.global.resolver.AuthAccountInfo;
-import com.dnd.spaced.global.resolver.AuthStore;
+import com.dnd.spaced.global.resolver.auth.AuthAccountInfo;
+import com.dnd.spaced.global.resolver.auth.AuthStore;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
