+++ spaced/src/main/java/com/dnd/spaced/global/resolver/auth/AuthAccountInfo.java
@@ -1,4 +1,4 @@
-package com.dnd.spaced.global.resolver;
+package com.dnd.spaced.global.resolver.auth;
 
 public record AuthAccountInfo(String email) {
 }
